poc/requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.23
httpx>=0.25
transitions>=0.9
pyyaml>=6.0
jsonpath-ng>=1.6
